feat: Add User entity class
This commit adds a new entity class `User` in the `andrehsvictor.kanshi.entity` package. The `User` class represents a user in the application and includes properties such as `id`, `username`, `password`, `email`, `likedAnimes`, `recommendations`, `createdAt`, and `updatedAt`. This class is necessary for storing and retrieving user data in the application's database.