feat: Add API endpoint to retrieve all recommendations for a specific anime
This commit adds a new API endpoint `/api/1.0/animes/{animeId}/recommendations/all` to the `RecommendationController` class. The endpoint allows users to retrieve all recommendations for a specific anime by providing the `animeId` as a path variable. This addition is necessary for providing a way to fetch all recommendations related to a particular anime.