feat: Add Dockerfile for building and running the application in a container
This commit adds a Dockerfile to the repository. The Dockerfile defines the build and runtime environment for the application. It uses the Maven image as the build stage and the Eclipse Temurin image as the runtime stage. The Dockerfile also sets the entrypoint to run the application using the generated JAR file. This addition enables easier deployment and distribution of the application using Docker containers.